resolve plugin:spa rendering
