+++ platform/src/bootstrap.js
@@ -10,6 +10,8 @@ global.sovereign = { logger }; // Make logger globally accessible (e.g., in Pris
 
 export async function bootstrap(manifest) {
   logger.info("🚀 Starting Sovereign platform...");
+  logger.info(`- Root directory: ${manifest.__rootdir}`);
+  logger.info(`- Plugin directory: ${manifest.__pluginsdir}`);
   const start = Date.now();
 
   try {
@@ -17,7 +19,6 @@ export async function bootstrap(manifest) {
 
     // Discovers and mounts all plugins under /__runtimeDir/plugins/*
     const __pluginsdir = path.resolve(manifest.__pluginsdir);
-    logger.info(`- Plugin directory: ${__pluginsdir}`);
     const extHost = await createExtHost(manifest, { pluginsDir: __pluginsdir });
 
     logger.info("- Initializing HTTP server...");
