+++ plugins/blog/views/blog/editor.html
@@ -224,7 +224,7 @@
             <a class="breadcrumb__link" href="/">Projects</a>
           </li>
           <li class="breadcrumb__item">
-            <a class="breadcrumb__link" href="/p/{{projectId}}">Blog</a>
+            <a class="breadcrumb__link" href="/blog/{{projectId}}">Blog</a>
           </li>
           <li class="breadcrumb__item" aria-current="page">
             <span class="breadcrumb__current">{{filename}}</span>
